Update product-package to accept decimals (#9)

+++ app/models/spree/active_shipping_product_decorator.rb
@@ -3,7 +3,7 @@ module Spree
     def self.prepended(base)
       base.has_many :product_packages, dependent: :destroy
 
-      base.accepts_nested_attributes_for :product_packages, allow_destroy: true, reject_if: lambda { |pp| (pp[:weight].blank? || (Integer(pp[:weight]) < 1)) }
+      base.accepts_nested_attributes_for :product_packages, allow_destroy: true, reject_if: lambda { |pp| (pp[:weight].blank? || (Float(pp[:weight]) < 1)) }
     end
   end
 end
